# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gas Delivery Management Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  name: string;
}

export interface LoginResponse {
  user: User;
  message?: string;
}

export interface Delivery {
  id: number;
  consumerNumber: string;
  customerName: string;
  mobileNumber: string;
  deliveryDate: string;
  nextEligibleDate: string;
  createdBy: number;
  createdByName?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDeliveryRequest {
  consumerNumber: string;
  customerName: string;
  deliveryDate: string;
}

export interface UpdateDeliveryRequest {
  consumerNumber?: string;
  customerName?: string;
  mobileNumber?: string;
  deliveryDate?: string;
}

export interface DeliveryListResponse {
  data: Delivery[];
  total: number;
  page: number;
  limit: number;
}

export interface Settings {
  /** Number of days a customer must wait before next gas delivery */
  waitingDays: number;
  updatedAt?: string;
}

export interface UpdateSettingsRequest {
  /** @minimum 1 */
  waitingDays: number;
}

export type GetDeliveriesParams = {
  /**
   * Search by consumer number, name, or mobile
   */
  search?: string;
  page?: number;
  limit?: number;
};