# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gas Delivery Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.number(),
    username: zod.string(),
    role: zod.enum(["admin", "user"]),
    name: zod.string(),
  }),
  message: zod.string().optional(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.enum(["admin", "user"]),
  name: zod.string(),
});

/**
 * @summary Get all delivery records
 */
export const getDeliveriesQueryPageDefault = 1;
export const getDeliveriesQueryLimitDefault = 50;

export const GetDeliveriesQueryParams = zod.object({
  search: zod.coerce
    .string()
    .optional()
    .describe("Search by consumer number, name, or mobile"),
  page: zod.coerce.number().default(getDeliveriesQueryPageDefault),
  limit: zod.coerce.number().default(getDeliveriesQueryLimitDefault),
});

export const GetDeliveriesResponse = zod.object({
  data: zod.array(
    zod.object({
      id: zod.number(),
      consumerNumber: zod.string(),
      customerName: zod.string(),
      mobileNumber: zod.string(),
      deliveryDate: zod.date(),
      nextEligibleDate: zod.date(),
      createdBy: zod.number(),
      createdByName: zod.string().optional(),
      createdAt: zod.date(),
      updatedAt: zod.date(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});

/**
 * @summary Create a delivery record
 */
export const CreateDeliveryBody = zod.object({
  consumerNumber: zod.string(),
  customerName: zod.string(),
  deliveryDate: zod.date(),
});

/**
 * @summary Get a delivery record by ID
 */
export const GetDeliveryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDeliveryResponse = zod.object({
  id: zod.number(),
  consumerNumber: zod.string(),
  customerName: zod.string(),
  mobileNumber: zod.string(),
  deliveryDate: zod.date(),
  nextEligibleDate: zod.date(),
  createdBy: zod.number(),
  createdByName: zod.string().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a delivery record
 */
export const UpdateDeliveryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDeliveryBody = zod.object({
  consumerNumber: zod.string().optional(),
  customerName: zod.string().optional(),
  mobileNumber: zod.string().optional(),
  deliveryDate: zod.date().optional(),
});

export const UpdateDeliveryResponse = zod.object({
  id: zod.number(),
  consumerNumber: zod.string(),
  customerName: zod.string(),
  mobileNumber: zod.string(),
  deliveryDate: zod.date(),
  nextEligibleDate: zod.date(),
  createdBy: zod.number(),
  createdByName: zod.string().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a delivery record
 */
export const DeleteDeliveryParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteDeliveryResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get portal settings
 */
export const GetSettingsResponse = zod.object({
  waitingDays: zod
    .number()
    .describe("Number of days a customer must wait before next gas delivery"),
  updatedAt: zod.date().optional(),
});

/**
 * @summary Update portal settings (admin only)
 */

export const UpdateSettingsBody = zod.object({
  waitingDays: zod.number().min(1),
});

export const UpdateSettingsResponse = zod.object({
  waitingDays: zod
    .number()
    .describe("Number of days a customer must wait before next gas delivery"),
  updatedAt: zod.date().optional(),
});