# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gas Delivery Management Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface LoginResponse {
  user: User;
  message?: string;
}